basic setup with boot, preload and game scene
