+++ src/scenes/bootScene.js
@@ -1,0 +1,17 @@
+import 'phaser';
+
+export default class bootScene extends Phaser.Scene {
+    constructor (){
+        super('boot');
+    }
+
+    preload ()
+    {
+        
+    }
+      
+    create ()
+    {
+        this.scene.start('preload');
+    }
+}
